+++ src/hike/widgets/command_line/widget.py
@@ -29,7 +29,13 @@ from textual_enhanced.commands import Quit
 from ...types import CommandHistory
 from .base_command import InputCommand
 from .change_directory import ChangeDirectoryCommand
-from .general import BookmarksCommand, ContentsCommand, HistoryCommand, QuitCommand
+from .general import (
+    BookmarksCommand,
+    ContentsCommand,
+    HistoryCommand,
+    LocalCommand,
+    QuitCommand,
+)
 from .open_directory import OpenDirectoryCommand
 from .open_file import OpenFileCommand
 from .open_from_forge import (
@@ -52,6 +58,7 @@ COMMANDS: Final[tuple[type[InputCommand], ...]] = (
     BookmarksCommand,
     ChangeDirectoryCommand,
     HistoryCommand,
+    LocalCommand,
     OpenFromBitbucket,
     OpenFromCodeberg,
     OpenFromGitHub,