+++ ChangeLog.md
@@ -18,6 +18,8 @@
   ([#24](https://github.com/davep/hike/pull/24))
 - Added `history` as a command that the command line understands.
   ([#24](https://github.com/davep/hike/pull/24))
+- Added `local` as a command that the command line understands.
+  ([#24](https://github.com/davep/hike/pull/24))
 
 ## v0.2.0